requirements.txt
asgiref==3.7.2
boto3==1.34.61
botocore==1.34.61
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
django-storages==1.14.2
idna==3.6
jmespath==1.0.1
mysqlclient==2.2.4
pillow==10.2.0
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.0.7
